+++ src/include/qtnode.hpp
@@ -7,7 +7,7 @@ public:
     QuadTreeNode(Quad boundary, int capacity);
     bool insert(Point* points, int point_index);
     void subdivide(Point* points);
-    void calculate_force_node(Point* points, int point_index, float& fx, float& fy, float softening_factor, float THETA, float G);
+    void calculate_force_node(Point* points, int point_index, double& fx, double& fy, double softening_factor, double THETA, double G);
     void clear();
     ~QuadTreeNode();
 
@@ -18,8 +18,8 @@ private:
     int num_stored_points_;
     bool divided_;
     QuadTreeNode* children_[4]; // TODO: std::unique_ptr
-    float total_mass_;
-    float center_x_, center_y_;
+    double total_mass_;
+    double center_x_, center_y_;
 };
 
 #endif
